Add stack layers for public load balancers and alternate hostnames
public_loadbalancer.yaml is a new stack that sits on top of an
infrastructure stack, using much of the same infrastructure but
creating a new, public facing load balancer. It exports the
outputs of the parent stack, except those relating to load
balancers, so any stacks created on top of this stakc will
attach to the new load balancer but use the same network and
kafka infrastructure.

alt_hostname.yaml is a new stack that attaches to a parent
infrastructure stack and attaches a new alb group with a
listener rule that watches for a different hostname. It points
a new DNS rule to the existing load balancer, and routes any
requests to that hostname to the target group created in the
stack. Like public_loadbalancer.yaml, it exposes outputs from
the parent stack, except those relating to the new alb group.

Finally, infrastructure_ext_kafka.yaml was updated to provide
a few outputs required by these stacks.