Adjust DatapointsToAlarm in Long alarms to prevent flapping
We observed that in some cases on our longer alarms, a server that
was in the healthy range would report an "ok" status, then a server
in the alarm state would report. The "Maximum" aggregator was intended
to take the edge off this, but sometimes the most recent period wouldn't
include the highest value, but it would be incorporated by the time
that datapoint was the second most recent data point. Looking at 28/30
instead of 30/30 data points makes sure the maximum always has data from
each server.