Sync control
Syncing from Kafka can impact the deltas for other servers, so we're
making some changes to do it more carefully. For initial syncs and
snapshots, we'll run with the avoid_leader flag, to pull from replicas
other than the leader. For snapshots, we're no longer running with
the flags that make initial syncs faster; we don't care if snapshots
take a few extra minutes compared with impacting consistency on
in-service instances.