Remove block checks
These block checks are taking a very long time to complete (much
longer than they seem like they should). We're going to remove them
at least temporarily while we figure out why.

They're not extremely necessary. We put them in because we saw issues
with blocks missing from the freezer, but that turned out to be an
issue with the way we wrote blocks into the freezer. We thought we'd
go ahead and implement these checks anyway under the expectation that
they would be cheap, but for non-obvious reasons they seem to be taking
an order of magnitude longer than the sync itself. Once we figure out
why, we may add them back in.