Large rework of Master,Fallback,Replica templates
Creates files needed on all servers to allow any server to run
as any other service. This should not be standard, but having
the ability could save a downtime incident.

This change also adds overlay functionality to the masters

This change also normalizes all the server setup to be the same
for every service, which will allow for more consistent server
behavior.

Remove the SIGINIT kill type for geth services -- causes issues