Volume refactor
This has a couple of volume related refactors.

First, we use gp3 instances across the board. Instead of starting
with io1 and downgrading to gp2, we start with gp3 and simply lower
the IOPS and throughput later.

Second, "full" sized replicas no longer have to be 'd' type instances.
They will start with an additional volume attached for the overlay,
which will be removed by a lambda function from d type instances. This
lets us run with all "r5x" type instances, which perform better for
our workload.