Stop using provisioned iops on startup
The performance benefits we used to see from provisioned iops on
cold snapshots seem to have disappeared. Now this is just a cost
with no measurable upside, so we're switching to GP3 end-to-end.

Also in this commit - GP3 volumes seem to have a momentary slowdown
after they are resized that we did not observe with gp2 volumes.
The master volume resize script will now only adjust the size of one
volume in each iteration, but every 15 minutes one will get resized
allowing us to resize before it becomes a problem, but staggered to
avoid all volumes hitting their slow period at once.