More robust pruning
This makes two changes in the name of better robustness for the
pruning process.

First, it allows us to use i3en.large instances for pruning. This
is necessary for mainnet, as even warm gp3 volumes have read latency
too high to prune the volume within our Kafka retention period (yes,
it's that bad). It makes it an option though, so testnets can use
r5.large / r5a.large for pruning to lower costs.

The other change is that the volume for pruning is considered to be
ephemeral - only used for pruning, snapshotted, then deleted. Previously
we had considered this to be the "next" snapshot volume, but tagging it
as such caused it to get resized to the current stack snapshot volume
size, and the resulting snapshot was sized accordingly. This way of
doing it does mean the first snapshot after pruning will a cold volume,
but it will be an appropriately sized volume.