Add a parameter to resume kafka from a given offset
When we prune in one region / account and want to copy a snapshot
to another environment, we typically end up pulling the entire history
of the Kafka topic, which takes a long time and is ugly.

It's possible to know an offset close to the range we want to resume
from, and our replica already supported a TopicName:Offset syntax, we
just couldn't use it from CloudFormation.

This lets us specify an offset for resumption in CloudFormation. When
a snapshot gets taken, that field will be emptied so it isn't a factor
in future restarts or snapshots.